MAYA-129059 - Add command to create stages.
Add an undoable command that creates a stage with a new layer.

The new command is a C++ implementation of the existing mayaUsd_createStageWithNewLayer.py Python script. The command enables the creation of stages from within the C++ code, which  will be required by future work (MAYA-128151) and might also be generally useful.

This commit also adds a Python binding for the stage creation command, which is used in mayaUsd_createStageWithNewLayer.py to replace the current implementation.